refactor + feat: toaster popup
