+++ src/app/layout.js
@@ -1,6 +1,8 @@
 import './globals.css'
 import { Nunito, Cabin_Sketch } from 'next/font/google'
 import { StudentDataProvider } from "@/StudentDataContext"
+import { ToastContainer } from 'react-toastify'
+import 'react-toastify/dist/ReactToastify.css'
 
 const nunito = Nunito({ 
   subsets: ["latin"], 
@@ -26,6 +28,10 @@ export default function RootLayout({ children }) {
       <StudentDataProvider>
         <body className={`${nunito.variable} ${cabinSketch.variable} font-nunito`}>
           {children}
+          <ToastContainer 
+            position="top-center"
+            pauseOnHover={false}
+          />
         </body>
       </StudentDataProvider>
     </html>
