+++ src/components/democlass/options/editTables/DeleteTableModal.jsx
@@ -1,12 +1,48 @@
-import React from 'react'
+import React, {useContext} from 'react'
 import { Dialog } from '@headlessui/react'
 import { AiOutlineClose } from "react-icons/ai"
+import StudentDataContext from "@/StudentDataContext"
+import { collection, updateDoc, doc } from 'firebase/firestore'
+import { db } from "../../../../utils/firebase"
 
-const DeleteTableModal = ( {checkDeleteTableModal, setCheckDeleteTableModal, deleteTable } ) => {
+const DeleteTableModal = ( {selectedTableName, openCheckDeleteTableModal, setOpenCheckDeleteTableModal, setOpenTableInfo} ) => {
+    const { studentData, setStudentData, userUid, userClassName } = useContext(StudentDataContext)
+
+    const deleteTable = async () => {
+        try {
+          // Reset demoStudentData tableData property back to default and deletes table from demoClass
+          const updatedStudentData = studentData.map((student) => {
+            if (student.tableData.tableName === selectedTableName) {
+              return {...student, tableData: {tableName: "", tablePoints: 0, isOnTable: false, selected: false}}
+            }
+            return student
+          })
+
+          // Set the updated student data to the state
+          setStudentData(updatedStudentData)
+
+          if(userUid && userClassName) {
+            // User is in their own class context (Firebase)
+            const classCollectionRef = collection(db, 'users', userUid, userClassName)
+            const classDocumentRef = doc(classCollectionRef, userUid)
+      
+            // Update the Firestore document with the updated studentData (resets studentData tableData property and deletes table from users class)
+            await updateDoc(classDocumentRef, {
+              studentData: updatedStudentData,
+            })
+          }
+
+        } catch (error) {
+          console.error('Error updating student information:', error)
+        }
+
+        setOpenCheckDeleteTableModal(false)
+        setOpenTableInfo(false)
+    }
   return (
     <Dialog
-        open={checkDeleteTableModal}
-        onClose={() => setCheckDeleteTableModal(false)}
+        open={openCheckDeleteTableModal}
+        onClose={() => setOpenCheckDeleteTableModal(false)}
         className="relative z-[100]"
         >
         {/* Backdrop */}
@@ -17,7 +53,7 @@ const DeleteTableModal = ( {checkDeleteTableModal, setCheckDeleteTableModal, del
             <Dialog.Panel className="flex flex-col p-5 w-full max-w-[550px] h-auto overflow-auto rounded-xl bg-modalBgClr">
             <div className="flex justify-between items-center border-b border-gray-400 pb-4">
                 <Dialog.Title className="font-bold text-lg">Delete this table group?</Dialog.Title>
-                <button onClick={() => setCheckDeleteTableModal(false)}>
+                <button onClick={() => setOpenCheckDeleteTableModal(false)}>
                 <AiOutlineClose
                     size={28}
                     className="bg-white text-secondaryTextClr hover:bg-buttonClr rounded-full hover:text-primaryTextClr p-1"
@@ -29,7 +65,7 @@ const DeleteTableModal = ( {checkDeleteTableModal, setCheckDeleteTableModal, del
                 <p className="text-sm font-bold text-red-500 text-center py-6">Are you sure you want to delete this table group? This can't be undone.</p>
                 <div className="flex justify-evenly items-center w-full border-t border-gray-400">
                 <button 
-                    onClick={() => setCheckDeleteTableModal(false)}
+                    onClick={() => setOpenCheckDeleteTableModal(false)}
                     className="bg-modalBgClr hover:bg-white rounded-xl py-2 px-3 text-buttonClr font-bold mt-4"
                     >
                     Cancel
